fix(elections): rename data_model module to prevent cross-plugin collision
The core loads plugin top-level modules by bare name on sys.path. Both
ledmatrix-elections and ledmatrix-flights shipped data_model.py, and elections'
providers/ subpackage imports it lazily during instantiation — after the loader
namespace-isolates bare names — so the import re-resolved via sys.path and bound
flights' data_model, failing to load (order-dependent).

Rename to election_data_model.py (plugin-unique) so the deferred import always
binds elections' own module regardless of load order. Relative imports aren't an
option: the loader execs the entry point with no package context.
