+++ plugins.json
@@ -846,7 +846,7 @@
       "last_updated": "2026-06-06",
       "verified": false,
       "screenshot": "",
-      "latest_version": "1.0.1"
+      "latest_version": "1.0.2"
     },
     {
       "id": "ledmatrix-dresden-departures",
