+++ plugins/ledmatrix-elections/election_data_model.py
